chore: move to output folder
