+++ .prettierignore
@@ -1,1 +1,2 @@
 __generated__/**
+output/**
