+++ src/vars.ts
@@ -1,0 +1,1 @@
+export const OUTPUT_DIR = './output';
